requirements.txt
Django==1.10.5
mysqlclient==1.3.9
django-debug-toolbar==1.6
djangorestframework